# docs/swagger.yaml
basePath: /
definitions:
  model.CommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      messgae:
        type: string
      photo_id:
        type: string
      updated_at:
        type: string
    type: object
  model.CreateSocialmediaRequest:
    properties:
      email:
        type: string
      name:
        type: string
      social_media_url:
        type: string
    required:
    - name
    - social_media_url
    type: object
  model.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  model.PhotoResponse:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: string
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  model.RegisterUserRequest:
    properties:
      age:
        minimum: 8
        type: integer
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  model.SocialmediaResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
    type: object
  model.UpdateCommentRequest:
    properties:
      message:
        type: string
      photo_id:
        type: string
    required:
    - message
    - photo_id
    type: object
  model.UpdatePhotoRequest:
    properties:
      caption:
        type: string
      email:
        type: string
      photo_url:
        type: string
      title:
        type: string
    required:
    - photo_url
    - title
    type: object
  model.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:5000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample API Example
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TEST API
  version: "2.0"
paths:
  /comment/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Comment
      parameters:
      - description: Comment Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Delete Comment
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: Find Comment
      parameters:
      - description: Comment Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Find Comment
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: Edit Comment
      parameters:
      - description: Update Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCommentRequest'
      - description: Comment Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Edit Comment
      tags:
      - comment
  /login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Login user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Login user
      tags:
      - users
  /photo:
    get:
      consumes:
      - application/json
      description: List Photo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PhotoResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Photo
      tags:
      - photo
  /photo/{id}:
    delete:
      consumes:
      - application/json
      description: Remove photo by ID
      parameters:
      - description: Photo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WebResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Remove photo by ID
      tags:
      - photo
  /photo/id:
    get:
      consumes:
      - application/json
      description: List Photo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PhotoResponse'
              type: object
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Photo
      tags:
      - photo
    put:
      consumes:
      - application/json
      description: Edit Photo
      parameters:
      - description: Update photo
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePhotoRequest'
      - description: Photo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PhotoResponse'
              type: object
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Edit Photo
      tags:
      - photo
  /register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: Register user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Register a new user
      tags:
      - users
  /socialmedia/:
    get:
      consumes:
      - application/json
      description: list socialmedia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SocialmediaResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: list socialmedia
      tags:
      - socialmedia
    post:
      consumes:
      - application/json
      description: Register socialmedia
      parameters:
      - description: Create Social Media
        in: body
        name: socialmedia
        required: true
        schema:
          $ref: '#/definitions/model.CreateSocialmediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Register a new socialmedia
      tags:
      - socialmedia
  /socialmedia/{id}:
    delete:
      consumes:
      - application/json
      description: Delete socialmedia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Delete Socialmedia
      tags:
      - socialmedia
    get:
      consumes:
      - application/json
      description: find socialmedia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.SocialmediaResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: find socialmedia
      tags:
      - socialmedia
    put:
      consumes:
      - application/json
      description: Edit socialmedia
      parameters:
      - description: Update socialmedia
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CreateSocialmediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.SocialmediaResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: edit socialmedia
      tags:
      - socialmedia
swagger: "2.0"